refactor: library.file 패키지를 library.reader 하위로 패키지로 이동
