+++ src/test/java/com/bookspot/batch/library/file/LibraryFileReaderConfigTest.java
@@ -1,11 +1,9 @@
 package com.bookspot.batch.library.file;
 
 import com.bookspot.batch.library.data.Library;
-import org.junit.jupiter.api.Test;
+import com.bookspot.batch.library.reader.file.LibraryFileReaderConfig;
 import org.springframework.batch.extensions.excel.poi.PoiItemReader;
 import org.springframework.batch.item.ExecutionContext;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class LibraryFileReaderConfigTest {
     LibraryFileReaderConfig config = new LibraryFileReaderConfig();
